🚚 Rename: 내 정보 조회 msw handler를 auth 핸들러와 파일 분리 (#274)
