+++ src/mocks/handlers/auth-handler.ts
@@ -1,6 +1,5 @@
 import { MSW_BASE_URL } from '@/constants/url-constants'
 import { http, HttpResponse, passthrough } from 'msw'
-import { userInformationMock } from '@/mocks/data/user-information-data'
 const ACCESS_TOKEN = `msw-access-token=access-token-test; Path=/; SameSite=Strict;`
 const REFRESH_TOKEN =
   'msw-refresh-token=refresh-token-test; Path=/; SameSite=Strict;'
@@ -41,20 +40,6 @@ const logout = http.post(`${MSW_BASE_URL}/auth/logout`, () => {
       headers: { 'Set-Cookie': `${REFRESH_TOKEN} Max-Age=0` },
     }
   )
-})
-
-// 내 정보 조회
-const getUserInformation = http.get(`${MSW_BASE_URL}/info`, ({ request }) => {
-  const header = request.headers.get('Authorization')
-  const hasBearerToken = header?.includes('Bearer')
-
-  if (!hasBearerToken) {
-    return HttpResponse.json(
-      { detail: 'Authentication required.' },
-      { status: 401 }
-    )
-  }
-  return HttpResponse.json(userInformationMock[0])
 })
 
 // 액세스 토큰 재발급
@@ -231,7 +216,6 @@ const passthroughPiscumPhotos = http.get(
 export const authHandlers = [
   login,
   logout,
-  getUserInformation,
   getRefreshToken,
   emailSendCode,
   emailVerify,
